requirements.txt
# python>=3.8
python-json-logger>=2.0
pydantic>=2.3.0
python-dotenv>=1.0.0
httpx>=0.24
pandas>=2.0
ijson>=3.2
tenacity>=8.2
aiofiles>=23.2
dask>=2023.5
simplejson>=3.19